Simplify screw insert creation code
Responsibility for creating and positioning shapes of screw insert
holes was scattered across multiple methods which made it hard to
maintain. In particular, dimensions of the screw insert shape had to
be passed to the `screw-insert` and `screw-placement` functions which
do not need this information, as they mostly do positioning work.
Screw insert shapes are now created directly in `screw-insert-hole`
and others. `screw-insert` is renamed to `screw-placement-common` to
better reflect the changes.